Move common hooks to respective folders
