+++ desktop/packages/mullvad-vpn/src/renderer/components/views/app-info/components/version-list-item/hooks/useShowAlert.tsx
@@ -1,4 +1,4 @@
-import { useVersionConsistent } from '../../../hooks';
+import { useVersionConsistent } from '../../../../../../redux/hooks';
 
 export const useShowAlert = () => {
   const { consistent } = useVersionConsistent();
