+++ desktop/packages/mullvad-vpn/src/renderer/components/views/app-info/hooks/index.ts
@@ -1,9 +1,1 @@
-export * from './useConnectionIsBlocked';
-export * from './usePushAppUpgrade';
-export * from './usePushChangelog';
-export * from './useSetttingsShowBetaReleases';
-export * from './useVersionConsistent';
-export * from './useVersionCurrent';
-export * from './useVersionIsBeta';
-export * from './useVersionSuggestedUpgrade';
 export * from './useShowUpdateAvailable';
