+++ desktop/packages/mullvad-vpn/src/renderer/components/views/app-info/hooks/useShowUpdateAvailable.tsx
@@ -1,4 +1,4 @@
-import { useVersionSuggestedUpgrade } from './useVersionSuggestedUpgrade';
+import { useVersionSuggestedUpgrade } from '../../../../redux/version/hooks/useVersionSuggestedUpgrade';
 
 export const useShowUpdateAvailable = () => {
   const { suggestedUpgrade } = useVersionSuggestedUpgrade();
