+++ desktop/packages/mullvad-vpn/src/renderer/history/hooks/index.ts
@@ -1,0 +1,2 @@
+export * from './usePushAppUpgrade';
+export * from './usePushChangelog';
