+++ desktop/packages/mullvad-vpn/src/renderer/redux/connection/hooks/index.ts
@@ -1,0 +1,1 @@
+export * from './useConnectionIsBlocked';
