+++ desktop/packages/mullvad-vpn/src/renderer/redux/connection/hooks/useConnectionIsBlocked.tsx
@@ -1,4 +1,4 @@
-import { useSelector } from '../../../../redux/store';
+import { useSelector } from '../../store';
 
 export const useConnectionIsBlocked = () => {
   return { isBlocked: useSelector((state) => state.connection.isBlocked) };
