+++ desktop/packages/mullvad-vpn/src/renderer/redux/version/hooks/index.ts
@@ -1,0 +1,4 @@
+export * from './useVersionConsistent';
+export * from './useVersionCurrent';
+export * from './useVersionIsBeta';
+export * from './useVersionSuggestedUpgrade';
