+++ desktop/packages/mullvad-vpn/src/renderer/redux/version/hooks/useVersionConsistent.tsx
@@ -1,4 +1,4 @@
-import { useSelector } from '../../../../redux/store';
+import { useSelector } from '../../store';
 
 export const useVersionConsistent = () => {
   return { consistent: useSelector((state) => state.version.consistent) };
