+++ desktop/packages/mullvad-vpn/src/renderer/redux/version/hooks/useVersionCurrent.tsx
@@ -1,4 +1,4 @@
-import { useSelector } from '../../../../redux/store';
+import { useSelector } from '../../store';
 
 export const useVersionCurrent = () => {
   return { current: useSelector((state) => state.version.current) };
