+++ desktop/packages/mullvad-vpn/src/renderer/redux/version/hooks/useVersionIsBeta.tsx
@@ -1,4 +1,4 @@
-import { useSelector } from '../../../../redux/store';
+import { useSelector } from '../../store';
 
 export const useVersionIsBeta = () => {
   return { isBeta: useSelector((state) => state.version.isBeta) };
