+++ desktop/packages/mullvad-vpn/src/renderer/redux/version/hooks/useVersionSuggestedUpgrade.tsx
@@ -1,4 +1,4 @@
-import { useSelector } from '../../../../redux/store';
+import { useSelector } from '../../store';
 
 export const useVersionSuggestedUpgrade = () => {
   return { suggestedUpgrade: useSelector((state) => state.version.suggestedUpgrade) };
